ShahilPortfolio/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.0.1",
    "wtforms>=3.2.1",
]
